Rename storyformatter to htmlformatter
